requirements.txt
aioredis==1.2.0
amqp==2.5.0
anyjson==0.3.3
asgi-redis==1.4.3
asgiref==3.1.4
asn1crypto==0.24.0
async-timeout==2.0.1
attrs==19.1.0
autobahn==19.7.2
Automat==0.7.0
billiard==3.6.0.0
celery==4.3.0
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
Django==2.2.3
django-decorator-include==2.1
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
kombu==4.6.3
msgpack==0.6.1
msgpack-python==0.5.6
psycopg2-binary==2.8.3
pycparser==2.19
PyHamcrest==1.9.0
pytz==2019.1
redis==3.3.3
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.1
txaio==18.8.1
vine==1.3.0
zope.interface==4.6.0
